requirements.txt
django>=1.9
django-celery>=3.1.15
django-registration>=2.0
stripe>=1.30
beautifulsoup4==4.4.1
requests==2.9.0
south
pyenchant
apscheduler>=3.0.5


